==> code/1.limpieza.R <==
            # -------------------------------#
            # Panel de tratados y Controles  #
            # -------------------------------#

# Creamos una base de datos preliminar de manzanas tratadas y manzanas de control
# A esta base se le agregan los datos simulados en el script de simulación
# La base de tratados/controles se crea con base en información de la primera línea del metro de Bogotá
# Los tratados salen del área de influencia de la primera línea, los controles se crean duplicando el tamaño del área de influencia en un buffer. 
# Algunas capas se crean en ArcGIS y se utilizan en este código. 

rm(list = ls())

library(sf)
library(dplyr)
library(purrr)
library(stringr)
library(data.table)
library(lwgeom)
library(tidyr)
library(spdep)
library(terra)


# ---------------------------
# Rutas y carga de archivos
# ---------------------------

setwd ("C:/Users/USUARIO/Documents/GitHub/TF_EU_STCH")

crs_bog <- 3116

# Sacamos el SHP del área de influencia de una caracterización de población hecha por la EMB (Empresa Metro de Bogotá)
# La caracterización poblacional está hecha con base en el censo de 2018
# Por ende, las capas que utilizamos como base para crear la base de datos simulada son de 2018. 
# Como la caracterización poblacional cuenta con coordenadas (lat, lon), utilizamos y cruzamos con un shapefile de precios comerciales promedio por manzana de 2018. 

poblacion <- read.csv("data/raw/20251204_carac_poblacional.csv", header = TRUE, stringsAsFactors = FALSE)

poblacion <- poblacion %>%
  mutate(longitud = gsub("'", "", longitud)) %>%
  mutate(longitud = as.numeric(longitud))

pob_sf <- st_as_sf(
  poblacion,
  coords = c("longitud", "latitud"),
  crs = 4326
)

pob_sf <- st_transform(pob_sf, crs_bog)

# Incluimos el límite de intervención (también definido en 2018)
interv <- st_read("data/raw/limite_intervencion/LIMITE_INTERVENCION.shp") %>%
  st_transform(crs_bog)

# ---------------------------
# Encontrar área de influencia en manzanas
# ---------------------------

# Leemos la capa a nivel manzana

precios_2018 <- st_read("data/raw/valor_ref_m_2018/Valor_Ref_M_2018.shp") %>% 
  rename_with(tolower) 

precios_2018 <-  st_transform(precios_2018, crs_bog)

# Join espacial para unir puntos con polígonos de manzana y filtrar por área de influencia (solamente manzanas con puntos)

pob_joined <- st_join(
  pob_sf,
  precios_2018,
  join = st_within,
  left = FALSE
)

# Nos aseguramos de que solo haya un dato de población por manzana, y si no se hace la suma. 

pob_agregada <- pob_joined %>%
  st_drop_geometry() %>%
  group_by(mancodigo) %>%
  summarise(
    pob_total = sum(Personas.por.manzana, na.rm = TRUE)
  )

# Filtramos aquellos códigos de manzana que tenían observaciones en el área de influencia. 

manzanas_2018 <- precios_2018 %>%
  filter(mancodigo %in% pob_agregada$mancodigo) %>%
  left_join(pob_agregada, by = "mancodigo")


# ---------------------------
# Crear contrafactual
# ---------------------------

# Creación de grupo control en ArcGIS: 
# Utilizamos la misma capa de manzanas en 2018 para crear el grupo de control para las manzanas del área de influencia
# Calculamos un buffer del mismo tamaño del área de influencia con respecto a la línea de intervención (buffer de 1.5 kilómetros aproximadamente)
# Utilizamos ArcGIS debido al nivel de detalle de las capas de manzanas y el tamaño de los archivos, que alentan los procesos en R. 

# Leer shapefiles
tratadas <- manzanas_2018 %>%   
  mutate(
  mancodigo = as.character(mancodigo),
  treatment = 1
)

controles <- st_read("data/intermediate/ArcGIS/manzanas_control/manzanas_control.shp") %>% 
  rename_with(tolower) %>%
  mutate(
  mancodigo = as.character(mancodigo),
  treatment = 0
  ) 

# Nos aseguramos de que no haya manzanas simultáneamente en ambos grupos. 

controles_filtrado <- controles %>% 
  filter(!(mancodigo %in% tratadas$mancodigo)) %>%
  st_transform(crs_bog)

manzanas_final <- bind_rows(tratadas, controles_filtrado)

manzanas_final <- manzanas_final %>% st_make_valid()

st_write(manzanas_final, "data/intermediate/R/manzanas_final.shp", delete_dsn = TRUE)

# ---------------------------
# Base final
# ---------------------------

# Inclusión de tratamiento escalonado en ArcGIS. 
# Dividimos la capa de manzanas tratadas y control (manzanas_final.shp) en 5 tramos. 
# Cada tramo empieza a ser tratado en un año entre 2021-2025. 
# Las divisiones se hacen de acuerdo con lo reportado por la EMB
# Todas las divisiones se realizan con respecto a avenidas principales

manzanas <- st_read("data/intermediate/ArcGIS/manzanas_stag/manzanas_stag.shp") %>% 
  rename_with(tolower)

# Nos aseguramos de que la variable de tratamiento escalonado sea NA para los Nunca tratados (grupo control). 

manzanas <- manzanas %>%
  mutate(treatment_ = if_else(treatment == 0, NA_character_, treatment_))

st_write(manzanas, "data/intermediate/R/manzanas.shp", delete_dsn = TRUE)
